django_project/requirements.txt
Django                   3.2.6
django-crispy-forms      1.12.0
django-phonenumber-field 5.2.0
